Create ~/concepts/metrics and MetricsCommonContext to begin reusing code between Model Metrics and Distributed Workloads
