+++ frontend/src/concepts/distributedWorkloads/DistributedWorkloadsContext.tsx
@@ -2,11 +2,7 @@ import * as React from 'react';
 import { Bullseye, Alert } from '@patternfly/react-core';
 import { ClusterQueueKind, WorkloadKind } from '~/k8sTypes';
 import { FetchStateObject } from '~/types';
-import {
-  DEFAULT_LIST_FETCH_STATE,
-  DEFAULT_VALUE_FETCH_STATE,
-  POLL_INTERVAL,
-} from '~/utilities/const';
+import { DEFAULT_LIST_FETCH_STATE, DEFAULT_VALUE_FETCH_STATE } from '~/utilities/const';
 import { SupportedArea, conditionalArea } from '~/concepts/areas';
 import useSyncPreferredProject from '~/concepts/projects/useSyncPreferredProject';
 import { ProjectsContext, byName } from '~/concepts/projects/ProjectsContext';
@@ -19,8 +15,8 @@ import {
   useDWWorkloadCurrentMetrics,
   useDWWorkloadTrendMetrics,
 } from '~/api';
-// TODO mturley these imports from ~/pages/modelServing/* should be moved somewhere page-agnostic
-import { TimeframeTitle } from '~/pages/modelServing/screens/types';
+import { RefreshIntervalValue } from '~/concepts/metrics/const';
+import { MetricsCommonContext } from '~/concepts/metrics/MetricsCommonContext';
 import useClusterQueues from './useClusterQueues';
 import useWorkloads from './useWorkloads';
 
@@ -31,10 +27,6 @@ type DistributedWorkloadsContextType = {
   workloadCurrentMetrics: DWWorkloadCurrentMetrics;
   workloadTrendMetrics: DWWorkloadTrendMetrics;
   refreshAllData: () => void;
-  refreshRate: number;
-  setRefreshRate: (interval: number) => void;
-  lastUpdateTime: number;
-  setLastUpdateTime: (time: number) => void;
   namespace?: string;
 };
 
@@ -72,10 +64,6 @@ export const DistributedWorkloadsContext = React.createContext<DistributedWorklo
     },
   },
   refreshAllData: () => undefined,
-  refreshRate: POLL_INTERVAL,
-  setRefreshRate: () => undefined,
-  lastUpdateTime: 0,
-  setLastUpdateTime: () => undefined,
 });
 
 export const DistributedWorkloadsContextProvider =
@@ -87,8 +75,10 @@ export const DistributedWorkloadsContextProvider =
     const project = projects.find(byName(namespace)) ?? null;
     useSyncPreferredProject(project);
 
-    const [refreshRate, setRefreshRate] = React.useState(POLL_INTERVAL);
-    const [lastUpdateTime, setLastUpdateTime] = React.useState<number>(Date.now());
+    const { currentTimeframe, currentRefreshInterval, lastUpdateTime, setLastUpdateTime } =
+      React.useContext(MetricsCommonContext);
+
+    const refreshRate = RefreshIntervalValue[currentRefreshInterval];
 
     // TODO mturley implement lazy loading, let the context consumers tell us what data they need and make the other ones throw a NotReadyError
 
@@ -97,9 +87,8 @@ export const DistributedWorkloadsContextProvider =
     const projectMetrics = useDWProjectMetrics(namespace, refreshRate);
     const workloadCurrentMetrics = useDWWorkloadCurrentMetrics(namespace, refreshRate);
 
-    // TODO mturley this timeframe param is a placeholder, wire it up to real timeframe selector state
     const workloadTrendMetrics = useDWWorkloadTrendMetrics(
-      TimeframeTitle.ONE_DAY,
+      currentTimeframe,
       lastUpdateTime,
       setLastUpdateTime,
       namespace,
@@ -148,10 +137,6 @@ export const DistributedWorkloadsContextProvider =
           workloadCurrentMetrics,
           workloadTrendMetrics,
           refreshAllData,
-          refreshRate,
-          setRefreshRate,
-          lastUpdateTime,
-          setLastUpdateTime,
           namespace,
         }}
       >
