+++ frontend/src/pages/modelServing/screens/metrics/performance/PerformanceTab.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import { PageSection, Stack, StackItem } from '@patternfly/react-core';
 import ModelGraphs from '~/pages/modelServing/screens/metrics/performance/ModelGraphs';
-import MetricsPageToolbar from '~/pages/modelServing/screens/metrics/MetricsPageToolbar';
+import MetricsPageToolbar from '~/concepts/metrics/MetricsPageToolbar';
 import { ModelMetricType } from '~/pages/modelServing/screens/metrics/ModelServingMetricsContext';
 import EnsureMetricsAvailable from '~/pages/modelServing/screens/metrics/EnsureMetricsAvailable';
