+++ frontend/src/pages/modelServing/screens/metrics/performance/ServerMetricsPage.tsx
@@ -1,6 +1,6 @@
 import { PageSection, Stack, StackItem } from '@patternfly/react-core';
 import React, { ReactElement } from 'react';
-import MetricsPageToolbar from '~/pages/modelServing/screens/metrics/MetricsPageToolbar';
+import MetricsPageToolbar from '~/concepts/metrics/MetricsPageToolbar';
 import ServerGraphs from '~/pages/modelServing/screens/metrics/performance/ServerGraphs';
 import { ServerMetricType } from '~/pages/modelServing/screens/metrics/ModelServingMetricsContext';
 import EnsureMetricsAvailable from '~/pages/modelServing/screens/metrics/EnsureMetricsAvailable';
