+++ frontend/src/pages/modelServing/screens/projects/ProjectServerMetricsWrapper.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import MetricsPage from '~/pages/modelServing/screens/metrics/MetricsPage';
+import { MetricsCommonContextProvider } from '~/concepts/metrics/MetricsCommonContext';
 import { ModelServingMetricsProvider } from '~/pages/modelServing/screens/metrics/ModelServingMetricsContext';
 import { getServerMetricsQueries } from '~/pages/modelServing/screens/metrics/utils';
 import { getProjectDisplayName } from '~/pages/projects/utils';
@@ -14,27 +15,29 @@ const ProjectServerMetricsWrapper: React.FC = () => (
       const projectDisplayName = getProjectDisplayName(currentProject);
       const serverName = getModelServerDisplayName(servingRuntime);
       return (
-        <ModelServingMetricsProvider
-          queries={queries}
-          type={PerformanceMetricType.SERVER}
-          namespace={projectDisplayName}
-        >
-          <MetricsPage
-            title={`${serverName} metrics`}
-            breadcrumbItems={[
-              { label: 'Data science projects', link: '/projects' },
-              {
-                label: projectDisplayName,
-                link: `/projects/${currentProject.metadata.name}`,
-              },
-              {
-                label: serverName,
-                isActive: true,
-              },
-            ]}
+        <MetricsCommonContextProvider>
+          <ModelServingMetricsProvider
+            queries={queries}
             type={PerformanceMetricType.SERVER}
-          />
-        </ModelServingMetricsProvider>
+            namespace={projectDisplayName}
+          >
+            <MetricsPage
+              title={`${serverName} metrics`}
+              breadcrumbItems={[
+                { label: 'Data science projects', link: '/projects' },
+                {
+                  label: projectDisplayName,
+                  link: `/projects/${currentProject.metadata.name}`,
+                },
+                {
+                  label: serverName,
+                  isActive: true,
+                },
+              ]}
+              type={PerformanceMetricType.SERVER}
+            />
+          </ModelServingMetricsProvider>
+        </MetricsCommonContextProvider>
       );
     }}
   </ProjectServerMetricsPathWrapper>
