+++ frontend/src/pages/modelServing/screens/types.ts
@@ -1,6 +1,7 @@
 import { SecretKind, ServingRuntimeKind } from '~/k8sTypes';
 import { EnvVariableDataEntry } from '~/pages/projects/types';
 import { ContainerResources } from '~/types';
+import { TimeframeStepType } from '~/concepts/metrics/types';
 import { ModelMetricType, ServerMetricType } from './metrics/ModelServingMetricsContext';
 
 export enum PerformanceMetricType {
@@ -14,38 +15,8 @@ export enum MetricType {
   BIAS = 'bias',
 }
 
-export enum TimeframeTitle {
-  ONE_HOUR = '1 hour',
-  ONE_DAY = '24 hours',
-  ONE_WEEK = '7 days',
-  ONE_MONTH = '30 days',
-  // UNLIMITED = 'Unlimited',
-}
-
-export type TimeframeTimeType = {
-  [key in TimeframeTitle]: number;
-};
-
-export type TimeframeStepType = TimeframeTimeType;
-
 export type QueryTimeframeStepType = {
   [key in ServerMetricType | ModelMetricType]: TimeframeStepType;
-};
-
-export enum RefreshIntervalTitle {
-  FIFTEEN_SECONDS = '15 seconds',
-  THIRTY_SECONDS = '30 seconds',
-  ONE_MINUTE = '1 minute',
-  FIVE_MINUTES = '5 minutes',
-  FIFTEEN_MINUTES = '15 minutes',
-  THIRTY_MINUTES = '30 minutes',
-  ONE_HOUR = '1 hour',
-  TWO_HOURS = '2 hours',
-  ONE_DAY = '1 day',
-}
-
-export type RefreshIntervalValueType = {
-  [key in RefreshIntervalTitle]: number;
 };
 
 export enum ServingRuntimeTableTabs {
